Add helper function for import to properly set project_id
